+++ apps/user/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 from apps.user.views.authentications import RegisterView, LoginView, LogoutView
-from apps.user.views.profiles import ChangePasswordView, MyProfileView
+from apps.user.views.profiles import ChangePasswordView, MyProfileView, UserListView
 
 urlpatterns = [
     path('auth/register/', RegisterView.as_view(), name='register'),
@@ -8,4 +8,6 @@ urlpatterns = [
     path('auth/logout/', LogoutView.as_view(), name='logout'),
     path('profile/', MyProfileView.as_view(), name='my-profile'),
     path('auth/change-password/', ChangePasswordView.as_view(), name='change-password'),
+
+    path('user/list/', UserListView.as_view(), name='user-list'),
 ]